% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_deleted_msg.R
\name{list_deleted_msg}
\alias{list_deleted_msg}
\title{List Deleted Messages}
\usage{
list_deleted_msg(
  msg_df,
  channel_ID_colname,
  message_ID_colname,
  target_entity = ""
)
}
\arguments{
\item{msg_df}{A data frame containing Telegram messages.}

\item{channel_ID_colname}{The column name of the channel ID in `msg_df`.}

\item{message_ID_colname}{The column name of the message ID in `msg_df`.}

\item{target_entity}{Channel ID of channel for which deleted messages should be searched. (optional). If none is chosen, all deleted messages in the dataset are searched.}
}
\value{
A data frame with deleted channel IDs and their corresponding deleted message IDs.
}
\description{
A function to list deleted messages based on the provided criteria.
}
