% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_ids_helper.R
\name{missing_ids_helper}
\alias{missing_ids_helper}
\title{Missing IDs Helper}
\usage{
missing_ids_helper(x, id_stats, msg_df, channel_ID, message_ID)
}
\arguments{
\item{x}{The index of the channel ID in `id_stats`.}

\item{id_stats}{A data frame containing channel IDs and maximum IDs.}

\item{msg_df}{A data frame containing messages.}

\item{channel_ID}{The column name of the channel ID in `msg_df`.}

\item{message_ID}{The column name of the message ID in `msg_df`.}
}
\value{
A data frame with missing channel IDs and their corresponding missing message IDs.
}
\description{
A helper function to identify missing IDs in a data frame.
}
