% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_deleted_msg.R
\name{find_deleted_msg}
\alias{find_deleted_msg}
\title{Find Deleted Messages}
\usage{
find_deleted_msg(
  corpus_df,
  deleted_msg_ids,
  deleted_channel_ids,
  fwd_from_colname,
  fwd_msg_colname,
  return_result = "all",
  date_colname = NA,
  impute = FALSE,
  message_ID_colname = NA,
  channel_ID_colname = NA
)
}
\arguments{
\item{corpus_df}{A data frame containing the corpus of messages.}

\item{deleted_msg_ids}{A vector of deleted message IDs.}

\item{deleted_channel_ids}{A vector of deleted channel IDs.}

\item{fwd_from_colname}{The column name stating which channel a message was forwarded from in `corpus_df`.}

\item{fwd_msg_colname}{The column name stating the original message IDs of the forwareded message in `corpus_df`.}

\item{return_result}{The type of result to return (all, latest, oldest).}

\item{date_colname}{The column name of the date variable in `corpus_df`.}

\item{impute}{Logical value indicating whether to impute messages.}

\item{message_ID_colname}{The column name of the message ID (optional, required for impute).}

\item{channel_ID_colname}{The column name of the channel ID (optional, required for impute).}
}
\value{
A data frame with the found deleted messages based on the specified criteria.
}
\description{
A function to find deleted messages based on the provided criteria.
}
